:art: Improve the code
