+++ Dockerfile
@@ -1,5 +1,9 @@
 FROM python:3.12.3
-
+ADD https://netfree.link/dl/unix-ca.sh /home/netfree-unix-ca.sh 
+RUN cat  /home/netfree-unix-ca.sh | sh
+ENV NODE_EXTRA_CA_CERTS=/etc/ca-bundle.crt
+ENV REQUESTS_CA_BUNDLE=/etc/ca-bundle.crt
+ENV SSL_CERT_FILE=/etc/ca-bundle.crt
 
 COPY . .
